+++ .eslintrc.cjs
@@ -52,7 +52,6 @@ module.exports = {
   },
   parser: '@babel/eslint-parser',
   parserOptions: {
-    parser: 'babel-eslint',
     ecmaVersion: 2018,
     sourceType: 'module',
     requireConfigFile: false,