+++ src/announcer/announcer.js
@@ -17,18 +17,122 @@
 
 import speechSynthesis from './speechSynthesis.js'
 
-let debounce
+let active = false
+let count = 0
+const queue = []
+let isProcessing = false
+let currentId = null
+let debounce = null
+
+const noopAnnouncement = {
+  then() {},
+  done() {},
+  cancel() {},
+  stop() {},
+}
+
+const enable = () => {
+  active = true
+}
+
+const disable = () => {
+  active = false
+}
+
+const toggle = (v) => {
+  active = v ? true : false
+}
 
 const speak = (message, politeness = 'off') => {
-  clearTimeout(debounce)
-  speechSynthesis.cancel()
-  // assertive messages get spoken immediately
-  if (politeness === 'assertive') {
-    speechSynthesis.speak({ value: message })
+  if (active === false) return noopAnnouncement
+
+  return addToQueue(message, politeness)
+}
+
+const pause = (delay) => {
+  if (active === false) return noopAnnouncement
+
+  return addToQueue(undefined, undefined, delay)
+}
+
+const addToQueue = (message, politeness, delay = false) => {
+  // keep track of the id so message can be canceled
+  const id = count++
+
+  // setup a promise to allow developer to chain functionality
+  // when specific utterances are done
+  let resolveFn
+  const done = new Promise((resolve) => {
+    resolveFn = resolve
+  })
+
+  // augment the promise with a cancel function
+  done.cancel = () => {
+    const index = queue.findIndex((item) => item.id === id)
+    if (index !== -1) queue.splice(index, 1)
+    resolveFn('canceled')
+  }
+
+  // augment the promise with a stop function
+  done.stop = () => {
+    if (id === currentId) {
+      speechSynthesis.cancel()
+      isProcessing = false
+      resolveFn('interupted')
+    }
+  }
+
+  // add message of pause
+  if (delay === false) {
+    politeness === 'assertive'
+      ? queue.unshift({ message, resolveFn, id })
+      : queue.push({ message, resolveFn, id })
   } else {
+    queue.push({ delay, resolveFn, id })
+  }
+
+  setTimeout(() => {
+    processQueue()
+  }, 100)
+
+  return done
+}
+
+const processQueue = async () => {
+  if (isProcessing === true || queue.length === 0) return
+  isProcessing = true
+
+  const { message, resolveFn, delay, id } = queue.shift()
+
+  currentId = id
+
+  if (delay) {
+    setTimeout(() => {
+      isProcessing = false
+      currentId = null
+      resolveFn('finished')
+      processQueue()
+    }, delay)
+  } else {
+    if (debounce !== null) clearTimeout(debounce)
+    // add some easing when speaking the messages to reduce stuttering
     debounce = setTimeout(() => {
-      speechSynthesis.speak({ value: message })
-    }, 400)
+      speechSynthesis
+        .speak({ message })
+        .then(() => {
+          isProcessing = false
+          currentId = null
+          resolveFn('finished')
+          processQueue()
+        })
+        .catch((e) => {
+          isProcessing = false
+          currentId = null
+          resolveFn(e.error)
+          processQueue()
+        })
+      debounce = null
+    }, 200)
   }
 }
 
@@ -40,9 +144,18 @@ const stop = () => {
   speechSynthesis.cancel()
 }
 
+const clear = () => {
+  queue.length = 0
+}
+
 export default {
   speak,
   polite,
   assertive,
   stop,
+  enable,
+  disable,
+  toggle,
+  clear,
+  pause,
 }